+++ crates/napi/src/react_compiler.rs
@@ -2,12 +2,12 @@ use std::{path::PathBuf, sync::Arc};
 
 use anyhow::Context as _;
 use napi::bindgen_prelude::*;
+use next_custom_transforms::react_compiler;
 use swc_core::{
     common::{SourceMap, GLOBALS},
     ecma::{
-        ast::{Callee, EsVersion, Expr, FnDecl, Program, ReturnStmt},
+        ast::EsVersion,
         parser::{parse_file_as_program, Syntax, TsSyntax},
-        visit::{Visit, VisitWith},
     },
 };
 
@@ -47,7 +47,7 @@ impl Task for CheckTask {
                 return Ok(false);
             }
 
-            Ok(is_required(&program))
+            Ok(react_compiler::is_required(&program))
         })
     }
 
@@ -63,59 +63,4 @@ pub fn is_react_compiler_required(
 ) -> AsyncTask<CheckTask> {
     let filename = PathBuf::from(filename);
     AsyncTask::with_optional_signal(CheckTask { filename }, signal)
-}
-
-fn is_required(program: &Program) -> bool {
-    let mut finder = Finder::default();
-    finder.visit_program(program);
-    finder.found
-}
-
-#[derive(Default)]
-struct Finder {
-    found: bool,
-
-    /// We are in a function that starts with a capital letter or it's a function that starts with
-    /// `use`
-    is_interested: bool,
-}
-
-impl Visit for Finder {
-    fn visit_callee(&mut self, node: &Callee) {
-        if self.is_interested {
-            if let Callee::Expr(e) = node {
-                if let Expr::Ident(c) = &**e {
-                    if c.sym.starts_with("use") {
-                        self.found = true;
-                        return;
-                    }
-                }
-            }
-        }
-
-        node.visit_children_with(self);
-    }
-
-    fn visit_fn_decl(&mut self, node: &FnDecl) {
-        let old = self.is_interested;
-        self.is_interested = node.ident.sym.starts_with("use")
-            || node.ident.sym.starts_with(|c: char| c.is_ascii_uppercase());
-
-        node.visit_children_with(self);
-
-        self.is_interested = old;
-    }
-
-    fn visit_return_stmt(&mut self, node: &ReturnStmt) {
-        if self.is_interested {
-            if let Some(Expr::JSXElement(..) | Expr::JSXEmpty(..) | Expr::JSXFragment(..)) =
-                node.arg.as_deref()
-            {
-                self.found = true;
-                return;
-            }
-        }
-
-        node.visit_children_with(self);
-    }
 }
