+++ crates/next-custom-transforms/src/lib.rs
@@ -39,6 +39,7 @@ use rustc_hash::FxHasher;
 
 pub mod chain_transforms;
 mod linter;
+pub mod react_compiler;
 pub mod transforms;
 
 type FxIndexMap<K, V> = IndexMap<K, V, BuildHasherDefault<FxHasher>>;
