+++ crates/next-custom-transforms/src/react_compiler.rs
@@ -1,0 +1,59 @@
+use swc_core::ecma::{
+    ast::{Callee, Expr, FnDecl, Program, ReturnStmt},
+    visit::{Visit, VisitWith},
+};
+
+pub fn is_required(program: &Program) -> bool {
+    let mut finder = Finder::default();
+    finder.visit_program(program);
+    finder.found
+}
+
+#[derive(Default)]
+struct Finder {
+    found: bool,
+
+    /// We are in a function that starts with a capital letter or it's a function that starts with
+    /// `use`
+    is_interested: bool,
+}
+
+impl Visit for Finder {
+    fn visit_callee(&mut self, node: &Callee) {
+        if self.is_interested {
+            if let Callee::Expr(e) = node {
+                if let Expr::Ident(c) = &**e {
+                    if c.sym.starts_with("use") {
+                        self.found = true;
+                        return;
+                    }
+                }
+            }
+        }
+
+        node.visit_children_with(self);
+    }
+
+    fn visit_fn_decl(&mut self, node: &FnDecl) {
+        let old = self.is_interested;
+        self.is_interested = node.ident.sym.starts_with("use")
+            || node.ident.sym.starts_with(|c: char| c.is_ascii_uppercase());
+
+        node.visit_children_with(self);
+
+        self.is_interested = old;
+    }
+
+    fn visit_return_stmt(&mut self, node: &ReturnStmt) {
+        if self.is_interested {
+            if let Some(Expr::JSXElement(..) | Expr::JSXEmpty(..) | Expr::JSXFragment(..)) =
+                node.arg.as_deref()
+            {
+                self.found = true;
+                return;
+            }
+        }
+
+        node.visit_children_with(self);
+    }
+}
